app.py
import streamlit as st
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Page configuration
st.set_page_config(
    page_title="Legal RAG System",
    page_icon="⚖️",
    layout="wide",
    initial_sidebar_state="expanded"
)

# Custom CSS
st.markdown("""
<style>
    .main-header {
        font-size: 2.5rem;
        color: #1f4e79;
        text-align: center;
        margin-bottom: 2rem;
    }
    .search-box {
        margin: 2rem 0;
    }
    .result-card {
        border: 1px solid #e0e0e0;
        border-radius: 8px;
        padding: 1rem;
        margin: 1rem 0;
        background-color: #f9f9f9;
    }
</style>
""", unsafe_allow_html=True)

def main():
    st.markdown('<h1 class="main-header">⚖️ Legal Document RAG System</h1>', unsafe_allow_html=True)
    
    with st.sidebar:
        st.header("🔍 Search Options")
        doc_types = st.multiselect(
            "Document Types",
            ["Code Juridique", "Arrétés", "Loi", "Circulaire", "Autres", "Décret", "Arrets"],
            default=["Code Juridique", "Loi"]
        )
        st.subheader("Date Range")
        col1, col2 = st.columns(2)
        with col1:
            start_year = st.number_input("From Year", value=2020, min_value=1970, max_value=2025)
        with col2:
            end_year = st.number_input("To Year", value=2024, min_value=1970, max_value=2025)
 
        max_results = st.slider("Max Results", min_value=5, max_value=50, value=10)

        if st.button("🗑️ Clear Search History"):
            st.session_state.messages = []
            st.rerun()
    if "messages" not in st.session_state:
        st.session_state.messages = []

    st.subheader("💬 Ask a Legal Question")

    for message in st.session_state.messages:
        with st.chat_message(message["role"]):
            st.markdown(message["content"])
    
    if prompt := st.chat_input("Enter your legal question..."):
        st.session_state.messages.append({"role": "user", "content": prompt})
        with st.chat_message("user"):
            st.markdown(prompt)
        
        with st.chat_message("assistant"):
            with st.spinner("Searching legal documents..."):
                # TODO: Replace with actual RAG implementation
                response = process_query(prompt, doc_types, start_year, end_year, max_results)
                st.markdown(response)
        st.session_state.messages.append({"role": "assistant", "content": response})

def process_query(query: str, doc_types: list, start_year: int, end_year: int, max_results: int) -> str:
    """
    Placeholder function for RAG processing
    TODO: Replace with actual vector search and LLM integration
    """
    return f"""
    **Query received:** "{query}"
    
    **Search Parameters:**
    - Document types: {', '.join(doc_types)}
    - Date range: {start_year} - {end_year}
    - Max results: {max_results}
    
    ---
    
    🚧 **This is a placeholder response.**
    
    Your RAG system will be integrated here to:
    1. Search your vector index for relevant documents
    2. Retrieve the most relevant chunks
    3. Generate a comprehensive answer using an LLM
    4. Provide source citations
    
    **Next step:** Implement the RAG service to connect to your existing index.
    """
def check_environment():
    """Check if required environment variables are set"""
    # required_vars = ["OPENAI_API_KEY"]  # Add your vector DB vars here
    # missing_vars = [var for var in required_vars if not os.getenv(var)]
    
    # if missing_vars:
    #     st.error(f"⚠️ Missing environment variables: {', '.join(missing_vars)}")
    #     st.info("Please create a `.env` file with the required variables. See `.env.example` for reference.")
    #     return False
    return True

if __name__ == "__main__":
    if check_environment():
        main()
    else:
        st.stop()