requirements.txt

streamlit==1.28.0
streamlit-chat==0.1.1
chromadb==0.4.18
openai==1.3.0
anthropic==0.8.0
langchain
langchain-openai
langchain_core
langchain_community
pandas
numpy
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
loguru==0.7.2
rapidfuzz